fix: ensure only allowed GuidedCapabilities are set
