+++ apps/ubuntu_bootstrap/lib/pages/storage/storage_model.dart
@@ -7,6 +7,7 @@ import 'package:safe_change_notifier/safe_change_notifier.dart';
 import 'package:subiquity_client/subiquity_client.dart';
 import 'package:ubuntu_bootstrap/pages.dart';
 import 'package:ubuntu_bootstrap/pages/storage/storage_page.dart';
+import 'package:ubuntu_bootstrap/pages/storage/storage_x.dart';
 import 'package:ubuntu_bootstrap/services.dart';
 import 'package:ubuntu_provision/ubuntu_provision.dart';
 
@@ -266,6 +267,13 @@ class StorageModel extends SafeChangeNotifier {
       StorageTypeManual() => null,
       null => null,
     };
+
+    if (guidedTarget?.allowed
+            .map((c) => c.clean())
+            .contains(guidedCapability) ==
+        false) {
+      _storage.guidedCapability = guidedTarget?.allowed.firstOrNull;
+    }
   }
 
   String? _resolvePartitionMethod() {
